simplify code for differential testing
